Move some helper functions to deploymentChecks
