+++ src/lib/deploymentChecks.ts
@@ -1,0 +1,43 @@
+// check to see if all replicas are ready for all deployments in the pepr-system namespace
+
+import { K8s, kind } from "kubernetes-fluent-client";
+import Log from "./logger";
+
+// returns true if all deployments are ready, false otherwise
+export async function checkDeploymentStatus(namespace: string) {
+  const deployments = await K8s(kind.Deployment).InNamespace(namespace).Get();
+  let status = false;
+  let readyCount = 0;
+
+  for (const deployment of deployments.items) {
+    const readyReplicas = deployment.status?.readyReplicas ? deployment.status?.readyReplicas : 0;
+    if (deployment.status?.readyReplicas !== deployment.spec?.replicas) {
+      Log.info(
+        `Waiting for deployment ${deployment.metadata?.name} rollout to finish: ${readyReplicas} of ${deployment.spec?.replicas} replicas are available`,
+      );
+    } else {
+      Log.info(
+        `Deployment ${deployment.metadata?.name} rolled out: ${readyReplicas} of ${deployment.spec?.replicas} replicas are available`,
+      );
+      readyCount++;
+    }
+  }
+  if (readyCount === deployments.items.length) {
+    status = true;
+  }
+  return status;
+}
+
+// wait for all deployments in the pepr-system namespace to be ready
+export async function namespaceDeploymentsReady(namespace: string = "pepr-system") {
+  Log.info(`Checking ${namespace} deployments status...`);
+  let ready = false;
+  while (!ready) {
+    ready = await checkDeploymentStatus(namespace);
+    if (ready) {
+      return ready;
+    }
+    await new Promise(resolve => setTimeout(resolve, 1000));
+  }
+  Log.info(`All ${namespace} deployments are ready`);
+}
